modified:   public/css/login.css 	modified:   public/css/nav.css 	modified:   public/css/results.css 	modified:   public/img/results/37.jpg 	modified:   public/js/results.js 	modified:   public/views/compass.php 	modified:   public/views/footer.php 	modified:   public/views/nav.php 	modified:   public/views/recommended.php 	modified:   public/views/results.php 	modified:   src/controllers/AppController.php 	modified:   src/controllers/DefaultController.php 	modified:   src/controllers/QuestionsController.php 	modified:   src/repository/Questions.php 	modified:   src/repository/UserRepository.php

+++ src/repository/Questions.php
@@ -38,7 +38,7 @@ class Questions extends Repository
         return $numberofquesitons['description'];
     }
 
-    public function getWidthValue(int $idq): ?int
+    public function getWidthValue(int $idq): ?float
     {
         $stmt = $this->database->connect()->prepare('
             SELECT value_w FROM questions
@@ -52,7 +52,7 @@ class Questions extends Repository
         return $numberofquesitons['value_w'];
     }
 
-    public function getHeightValue(int $idq): ?int
+    public function getHeightValue(int $idq): ?float
     {
         $stmt = $this->database->connect()->prepare('
             SELECT value_h FROM questions
